+++ tests/utils.rs
@@ -20,15 +20,15 @@ const TIMEOUT_S: f64 = 0.1;
 #[cfg(target_os="windows")]
 const TIMEOUT_S: f64 = 3.0; // windows can take a while
 
-pub fn recv_events(rx: &Receiver<Event>) ->  Vec<(PathBuf, Op)> {
+pub fn recv_events(rx: &Receiver<Event>) ->  Vec<(PathBuf, Op, Option<u32>)> {
     let deadline = time::precise_time_s() + TIMEOUT_S;
 
-    let mut evs: Vec<(PathBuf, Op)> = Vec::new();
+    let mut evs = Vec::new();
 
     while time::precise_time_s() < deadline {
         match rx.try_recv() {
-            Ok(Event{path: Some(path), op: Ok(op)}) => {
-                evs.push((path, op));
+            Ok(Event{path: Some(path), op: Ok(op), cookie}) => {
+                evs.push((path, op, cookie));
             },
             Ok(Event{path: None, ..})  => (),
             Ok(Event{op: Err(e), ..}) => panic!("unexpected event err: {:?}", e),
@@ -42,29 +42,46 @@ pub fn recv_events(rx: &Receiver<Event>) ->  Vec<(PathBuf, Op)> {
 // FSEvent tends to emit events multiple times and aggregate events,
 // so just check that all expected events arrive for each path,
 // and make sure the paths are in the correct order
-#[allow(dead_code)]
-pub fn inflate_events(input: Vec<(PathBuf, Op)>) -> Vec<(PathBuf, Op)> {
+pub fn inflate_events(input: Vec<(PathBuf, Op, Option<u32>)>) -> Vec<(PathBuf, Op, Option<u32>)> {
     let mut output = Vec::new();
     let mut path = None;
     let mut ops = Op::empty();
-    for (e_p, e_op) in input {
+    let mut cookie = None;
+    for (e_p, e_o, e_c) in input {
         let p = match path {
             Some(p) => p,
             None => e_p.clone()
         };
-        if p == e_p {
-            // ops |= e_op;
-            ops = Op::from_bits_truncate(ops.bits() | e_op.bits());
+        let c = match cookie {
+            Some(c) => Some(c),
+            None => e_c
+        };
+        if p == e_p && c == e_c {
+            // ops |= e_o;
+            ops = Op::from_bits_truncate(ops.bits() | e_o.bits());
         } else {
-            output.push((p, ops));
-            ops = e_op;
+            output.push((p, ops, cookie));
+            ops = e_o;
         }
         path = Some(e_p);
+        cookie = e_c;
     }
     if let Some(p) = path {
-        output.push((p, ops));
+        output.push((p, ops, cookie));
     }
     output
+}
+
+pub fn extract_cookies(events: &Vec<(PathBuf, Op, Option<u32>)>) -> Vec<u32> {
+    let mut cookies = Vec::new();
+    for &(_, _, e_c) in events {
+        if let Some(cookie) = e_c {
+            if !cookies.contains(&cookie) {
+                cookies.push(cookie);
+            }
+        }
+    }
+    cookies
 }
 
 // Sleep for `duration` in milliseconds