+++ tests/watcher.rs
@@ -10,7 +10,6 @@ use std::sync::mpsc;
 use tempdir::TempDir;
 use std::thread;
 use std::env;
-use std::path::Path;
 
 use utils::*;
 
@@ -88,7 +87,7 @@ fn watch_relative() {
 
         env::set_current_dir(tdir.path()).expect("failed to change working directory");
 
-        let (tx, rx) = mpsc::channel();
+        let (tx, _) = mpsc::channel();
         let mut watcher: RecommendedWatcher = Watcher::new(tx).expect("failed to create recommended watcher");
         watcher.watch("dir1", RecursiveMode::Recursive).expect("failed to watch directory");
     }
@@ -98,7 +97,7 @@ fn watch_relative() {
 
         env::set_current_dir(tdir.path()).expect("failed to change working directory");
 
-        let (tx, rx) = mpsc::channel();
+        let (tx, _) = mpsc::channel();
         let mut watcher: RecommendedWatcher = Watcher::new(tx).expect("failed to create recommended watcher");
         watcher.watch("file1", RecursiveMode::Recursive).expect("failed to watch file");
     }
@@ -139,7 +138,7 @@ fn watch_recursive_create_directory() {
     let actual = if cfg!(target_os="windows") {
         // Windows may sneak a write event in there
         let mut events = recv_events(&rx);
-        events.retain(|&(_, op)| op != op::WRITE);
+        events.retain(|&(_, op, _)| op != op::WRITE);
         events
     } else if cfg!(target_os="macos") {
         inflate_events(recv_events(&rx))
@@ -148,13 +147,12 @@ fn watch_recursive_create_directory() {
     };
 
     assert_eq!(actual, vec![
-        (tdir.mkpath("dir1"), op::CREATE),
-        (tdir.mkpath("dir1/file1"), op::CREATE)
+        (tdir.mkpath("dir1"), op::CREATE, None),
+        (tdir.mkpath("dir1/file1"), op::CREATE, None)
     ]);
 }
 
 #[test]
-#[ignore]
 fn watch_recursive_move() {
     let tdir = TempDir::new("temp_dir").expect("failed to create temporary directory");
 
@@ -182,7 +180,7 @@ fn watch_recursive_move() {
     let actual = if cfg!(target_os="windows") {
         // Windows may sneak a write event in there
         let mut events = recv_events(&rx);
-        events.retain(|&(_, op)| op != op::WRITE);
+        events.retain(|&(_, op, _)| op != op::WRITE);
         events
     } else if cfg!(target_os="macos") {
         inflate_events(recv_events(&rx))
@@ -190,27 +188,23 @@ fn watch_recursive_move() {
         recv_events(&rx)
     };
 
-    if cfg!(target_os="windows") {
+    if cfg!(target_os="macos") {
+        let cookies = extract_cookies(&actual);
+        assert_eq!(cookies.len(), 1);
         assert_eq!(actual, vec![
-            (tdir.mkpath("dir1a/file1"), op::CREATE),
-            (tdir.mkpath("dir1a"), op::RENAME),
-            (tdir.mkpath("dir1b"), op::RENAME), // should be a create
-            (tdir.mkpath("dir1b/file2"), op::CREATE)
-        ]);
-        panic!("move_to should be translated to create");
-    } else if cfg!(target_os="macos") {
-        assert_eq!(actual, vec![
-            (tdir.mkpath("dir1a/file1"), op::CREATE),
-            (tdir.mkpath("dir1a"), op::CREATE | op::RENAME), // excessive create event
-            (tdir.mkpath("dir1b"), op::RENAME), // should be a create
-            (tdir.mkpath("dir1b/file2"), op::CREATE)
+            (tdir.mkpath("dir1a/file1"), op::CREATE, None),
+            (tdir.mkpath("dir1a"), op::CREATE | op::RENAME, Some(cookies[0])), // excessive create event
+            (tdir.mkpath("dir1b"), op::RENAME, Some(cookies[0])),
+            (tdir.mkpath("dir1b/file2"), op::CREATE, None)
         ]);
     } else {
+        let cookies = extract_cookies(&actual);
+        assert_eq!(cookies.len(), 1);
         assert_eq!(actual, vec![
-            (tdir.mkpath("dir1a/file1"), op::CREATE),
-            (tdir.mkpath("dir1a"), op::RENAME),
-            (tdir.mkpath("dir1b"), op::CREATE),
-            (tdir.mkpath("dir1b/file2"), op::CREATE)
+            (tdir.mkpath("dir1a/file1"), op::CREATE, None),
+            (tdir.mkpath("dir1a"), op::RENAME, Some(cookies[0])),
+            (tdir.mkpath("dir1b"), op::RENAME, Some(cookies[0])),
+            (tdir.mkpath("dir1b/file2"), op::CREATE, None)
         ]);
     }
 }
@@ -244,7 +238,7 @@ fn watch_recursive_move_in() {
     let actual = if cfg!(target_os="windows") {
         // Windows may sneak a write event in there
         let mut events = recv_events(&rx);
-        events.retain(|&(_, op)| op != op::WRITE);
+        events.retain(|&(_, op, _)| op != op::WRITE);
         events
     } else if cfg!(target_os="macos") {
         inflate_events(recv_events(&rx))
@@ -254,19 +248,20 @@ fn watch_recursive_move_in() {
 
     if cfg!(target_os="macos") {
         assert_eq!(actual, vec![
-            (tdir.mkpath("watch_dir/dir1b"), op::RENAME), // fsevents interprets a move_to as a rename event
-            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE)
+            (tdir.mkpath("watch_dir/dir1b"), op::RENAME, None), // fsevent interprets a move_to as a rename event
+            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE, None)
         ]);
         panic!("move event should be a create event");
     } else {
         assert_eq!(actual, vec![
-            (tdir.mkpath("watch_dir/dir1b"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE)
+            (tdir.mkpath("watch_dir/dir1b"), op::CREATE, None),
+            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE, None)
         ]);
     }
 }
 
 #[test]
+#[ignore]
 fn watch_recursive_move_out() {
     let tdir = TempDir::new("temp_dir").expect("failed to create temporary directory");
 
@@ -294,7 +289,7 @@ fn watch_recursive_move_out() {
     let actual = if cfg!(target_os="windows") {
         // Windows may sneak a write event in there
         let mut events = recv_events(&rx);
-        events.retain(|&(_, op)| op != op::WRITE);
+        events.retain(|&(_, op, _)| op != op::WRITE);
         events
     } else if cfg!(target_os="macos") {
         inflate_events(recv_events(&rx))
@@ -302,20 +297,16 @@ fn watch_recursive_move_out() {
         recv_events(&rx)
     };
 
-    if cfg!(target_os="windows") {
+    if cfg!(target_os="macos") {
         assert_eq!(actual, vec![
-            (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1a"), op::REMOVE) // windows interprets a move out of the watched directory as a remove
+            (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::CREATE, None),
+            (tdir.mkpath("watch_dir/dir1a"), op::CREATE | op::RENAME, None) // excessive create event, fsevent interprets a move_out as a rename event
         ]);
-    } else if cfg!(target_os="macos") {
-        assert_eq!(actual, vec![
-            (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1a"), op::CREATE | op::RENAME) // excessive create event
-        ]);
+        panic!("move event should be a remove event");
     } else {
         assert_eq!(actual, vec![
-            (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1a"), op::RENAME)
+            (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::CREATE, None),
+            (tdir.mkpath("watch_dir/dir1a"), op::REMOVE, None),
         ]);
     }
 }
@@ -349,8 +340,8 @@ fn watch_nonrecursive() {
     ]);
 
     assert_eq!(recv_events(&rx), vec![
-        (tdir.mkpath("dir2"), op::CREATE),
-        (tdir.mkpath("file0"), op::CREATE)
+        (tdir.mkpath("dir2"), op::CREATE, None),
+        (tdir.mkpath("file0"), op::CREATE, None)
     ]);
 }
 
@@ -379,11 +370,11 @@ fn watch_file() {
 
     if cfg!(target_os="macos") {
         assert_eq!(recv_events(&rx), vec![
-            (tdir.mkpath("file1"), op::CREATE | op::WRITE)
+            (tdir.mkpath("file1"), op::CREATE | op::WRITE, None) // excessive write create
         ]);
     } else {
         assert_eq!(recv_events(&rx), vec![
-            (tdir.mkpath("file1"), op::WRITE)
+            (tdir.mkpath("file1"), op::WRITE, None)
         ]);
     }
 }
@@ -410,9 +401,9 @@ fn poll_watch_recursive_create_directory() {
     actual.sort_by(|a, b| a.0.cmp(&b.0));
 
     assert_eq!(actual, vec![
-        (tdir.mkpath("."), op::WRITE), // parent directory gets modified
-        (tdir.mkpath("dir1"), op::CREATE),
-        (tdir.mkpath("dir1/file1"), op::CREATE)
+        (tdir.mkpath("."), op::WRITE, None), // parent directory gets modified
+        (tdir.mkpath("dir1"), op::CREATE, None),
+        (tdir.mkpath("dir1/file1"), op::CREATE, None)
     ]);
 }
 
@@ -436,8 +427,8 @@ fn poll_watch_recursive_move() {
     actual.sort_by(|a, b| a.0.cmp(&b.0));
 
     assert_eq!(actual, vec![
-        (tdir.mkpath("dir1a"), op::WRITE), // parent directory gets modified
-        (tdir.mkpath("dir1a/file1"), op::CREATE),
+        (tdir.mkpath("dir1a"), op::WRITE, None), // parent directory gets modified
+        (tdir.mkpath("dir1a/file1"), op::CREATE, None),
     ]);
 
     sleep(1100); // PollWatcher has only a resolution of 1 second
@@ -450,22 +441,22 @@ fn poll_watch_recursive_move() {
 
     if cfg!(target_os="windows") {
         assert_eq!(actual, vec![
-            (tdir.mkpath("."), op::WRITE), // parent directory gets modified
-            (tdir.mkpath("dir1a"), op::REMOVE),
-            (tdir.mkpath("dir1a/file1"), op::REMOVE),
-            (tdir.mkpath("dir1b"), op::CREATE),
-            (tdir.mkpath("dir1b"), op::WRITE), // parent directory gets modified
-            (tdir.mkpath("dir1b/file1"), op::CREATE),
-            (tdir.mkpath("dir1b/file2"), op::CREATE),
+            (tdir.mkpath("."), op::WRITE, None), // parent directory gets modified
+            (tdir.mkpath("dir1a"), op::REMOVE, None),
+            (tdir.mkpath("dir1a/file1"), op::REMOVE, None),
+            (tdir.mkpath("dir1b"), op::CREATE, None),
+            (tdir.mkpath("dir1b"), op::WRITE, None), // parent directory gets modified
+            (tdir.mkpath("dir1b/file1"), op::CREATE, None),
+            (tdir.mkpath("dir1b/file2"), op::CREATE, None),
         ]);
     } else {
         assert_eq!(actual, vec![
-            (tdir.mkpath("."), op::WRITE), // parent directory gets modified
-            (tdir.mkpath("dir1a"), op::REMOVE),
-            (tdir.mkpath("dir1a/file1"), op::REMOVE),
-            (tdir.mkpath("dir1b"), op::CREATE),
-            (tdir.mkpath("dir1b/file1"), op::CREATE),
-            (tdir.mkpath("dir1b/file2"), op::CREATE),
+            (tdir.mkpath("."), op::WRITE, None), // parent directory gets modified
+            (tdir.mkpath("dir1a"), op::REMOVE, None),
+            (tdir.mkpath("dir1a/file1"), op::REMOVE, None),
+            (tdir.mkpath("dir1b"), op::CREATE, None),
+            (tdir.mkpath("dir1b/file1"), op::CREATE, None),
+            (tdir.mkpath("dir1b/file2"), op::CREATE, None),
         ]);
     }
 }
@@ -493,18 +484,18 @@ fn poll_watch_recursive_move_in() {
 
     if cfg!(target_os="windows") {
         assert_eq!(actual, vec![
-            (tdir.mkpath("watch_dir"), op::WRITE), // parent directory gets modified
-            (tdir.mkpath("watch_dir/dir1b"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1b/dir1"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1b/dir1"), op::WRITE), // extra write event
-            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE),
+            (tdir.mkpath("watch_dir"), op::WRITE, None), // parent directory gets modified
+            (tdir.mkpath("watch_dir/dir1b"), op::CREATE, None),
+            (tdir.mkpath("watch_dir/dir1b/dir1"), op::CREATE, None),
+            (tdir.mkpath("watch_dir/dir1b/dir1"), op::WRITE, None), // extra write event
+            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE, None),
         ]);
     } else {
         assert_eq!(actual, vec![
-            (tdir.mkpath("watch_dir"), op::WRITE), // parent directory gets modified
-            (tdir.mkpath("watch_dir/dir1b"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1b/dir1"), op::CREATE),
-            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE),
+            (tdir.mkpath("watch_dir"), op::WRITE, None), // parent directory gets modified
+            (tdir.mkpath("watch_dir/dir1b"), op::CREATE, None),
+            (tdir.mkpath("watch_dir/dir1b/dir1"), op::CREATE, None),
+            (tdir.mkpath("watch_dir/dir1b/dir1/file1"), op::CREATE, None),
         ]);
     }
 }
@@ -529,8 +520,8 @@ fn poll_watch_recursive_move_out() {
     actual.sort_by(|a, b| a.0.cmp(&b.0));
 
     assert_eq!(actual, vec![
-        (tdir.mkpath("watch_dir/dir1a/dir1"), op::WRITE), // parent directory gets modified
-        (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::CREATE),
+        (tdir.mkpath("watch_dir/dir1a/dir1"), op::WRITE, None), // parent directory gets modified
+        (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::CREATE, None),
     ]);
 
     sleep(1100); // PollWatcher has only a resolution of 1 second
@@ -542,10 +533,10 @@ fn poll_watch_recursive_move_out() {
     actual.sort_by(|a, b| a.0.cmp(&b.0));
 
     assert_eq!(actual, vec![
-        (tdir.mkpath("watch_dir"), op::WRITE), // parent directory gets modified
-        (tdir.mkpath("watch_dir/dir1a"), op::REMOVE),
-        (tdir.mkpath("watch_dir/dir1a/dir1"), op::REMOVE),
-        (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::REMOVE),
+        (tdir.mkpath("watch_dir"), op::WRITE, None), // parent directory gets modified
+        (tdir.mkpath("watch_dir/dir1a"), op::REMOVE, None),
+        (tdir.mkpath("watch_dir/dir1a/dir1"), op::REMOVE, None),
+        (tdir.mkpath("watch_dir/dir1a/dir1/file1"), op::REMOVE, None),
     ]);
 }
 
@@ -573,10 +564,10 @@ fn poll_watch_nonrecursive() {
     actual.sort_by(|a, b| a.0.cmp(&b.0));
 
     assert_eq!(actual, vec![
-        (tdir.mkpath("."), op::WRITE), // parent directory gets modified
-        (tdir.mkpath("dir1"), op::WRITE), // parent directory gets modified
-        (tdir.mkpath("dir2"), op::CREATE),
-        (tdir.mkpath("file1"), op::CREATE),
+        (tdir.mkpath("."), op::WRITE, None), // parent directory gets modified
+        (tdir.mkpath("dir1"), op::WRITE, None), // parent directory gets modified
+        (tdir.mkpath("dir2"), op::CREATE, None),
+        (tdir.mkpath("file1"), op::CREATE, None),
     ]);
 }
 
@@ -598,6 +589,6 @@ fn poll_watch_file() {
     tdir.create("file2");
 
     assert_eq!(recv_events(&rx), vec![
-        (tdir.mkpath("file1"), op::WRITE)
+        (tdir.mkpath("file1"), op::WRITE, None)
     ]);
 }