+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 #### Breaking Changes
 
 * Add RecursiveMode switch to Watcher::watch(..) (closes [#60](https://github.com/passcod/rsnotify/issues/60), [#61](https://github.com/passcod/rsnotify/issues/61))
+* Track move events using a cookie
 
 
 ### v2.6.2